tutorials/mobility/requirements.txt
glassflow
requests
python-dotenv
Faker==23.3.0
shapely==2.0.3